Move LFB migration to LastFinalizedKeyValueStorage
