+++ block-storage/src/main/scala/coop/rchain/blockstorage/dag/BlockDagKeyValueStorage.scala
@@ -1,6 +1,5 @@
 package coop.rchain.blockstorage.dag
 
-import cats.Show
 import cats.effect.concurrent.Semaphore
 import cats.effect.{Concurrent, Sync}
 import cats.syntax.all._
@@ -10,7 +9,7 @@ import coop.rchain.blockstorage.dag.BlockDagStorage.DeployId
 import coop.rchain.blockstorage.dag.BlockMetadataStore.BlockMetadataStore
 import coop.rchain.blockstorage.dag.EquivocationTrackerStore.EquivocationTrackerStore
 import coop.rchain.blockstorage.dag.codecs._
-import coop.rchain.blockstorage.finality.LastFinalizedStorage
+import coop.rchain.blockstorage.syntax._
 import coop.rchain.blockstorage.util.BlockMessageUtil._
 import coop.rchain.casper.PrettyPrinter
 import coop.rchain.casper.protocol.BlockMessage
@@ -21,8 +20,6 @@ import coop.rchain.models.EquivocationRecord.SequenceNumber
 import coop.rchain.models.Validator.Validator
 import coop.rchain.models.{BlockHash, BlockMetadata, EquivocationRecord, Validator}
 import coop.rchain.shared.syntax._
-import coop.rchain.blockstorage.syntax._
-import coop.rchain.dag.DagOps
 import coop.rchain.shared.{Log, LogSource}
 import coop.rchain.store.{KeyValueStoreManager, KeyValueTypedStore}
 
@@ -340,66 +337,4 @@ object BlockDagKeyValueStorage {
       stores.invalidBlocks,
       stores.equivocations
     )
-
-  def migrateLfb[F[_]: Sync: Log](
-      lastFinalizedStorage: LastFinalizedStorage[F],
-      kvm: KeyValueStoreManager[F]
-  ): F[Unit] = {
-    val errNoLfbInStorage   = "No LFB in LastFinalizedStorage when attempting migration."
-    val errNoMetadataForLfb = "No metadata found for LFB when attempting migration."
-
-    for {
-      blockMetadataDb <- kvm.database[BlockHash, BlockMetadata](
-                          "block-metadata",
-                          codecBlockHash,
-                          codecBlockMetadata
-                        )
-      // record LFB
-      lfb  <- lastFinalizedStorage.get() >>= (_.liftTo(new Exception(errNoLfbInStorage)))
-      curV <- blockMetadataDb.get(lfb).flatMap(_.liftTo[F](new Exception(errNoMetadataForLfb)))
-      _    <- blockMetadataDb.put(lfb, curV.copy(directlyFinalized = true, finalized = true))
-      blocksInfoMap <- blockMetadataDb
-                        .collect {
-                          case (hash, metaData) =>
-                            (hash, BlockMetadataStore.blockMetadataToInfo(metaData()))
-                        }
-                        .map(_.toMap)
-      _ <- Log[F].info("Migration of LFB done.")
-
-      // record finalized blocks
-      finalizedBlockSet = DagOps
-        .bfTraverse(List(lfb)) { bh =>
-          blocksInfoMap.get(bh).map(_.parents.toList).getOrElse(List.empty)
-        }
-        .toList
-
-      processChunk = (chunk: List[BlockHash]) => {
-        implicit val s = new Show[BlockHash] {
-          override def show(t: BlockHash): String = PrettyPrinter.buildString(t)
-        }
-
-        for {
-          curVs <- blockMetadataDb.getUnsafeBatch(chunk)
-          // WARNING: migration should be done before block merge, as it assumes all blocks are directly finalized.
-          newVs = curVs.map(_.copy(directlyFinalized = true, finalized = true))
-          _     <- blockMetadataDb.put(newVs.map(v => (v.blockHash, v)))
-        } yield ()
-      }
-
-      chunkSize = 10000L
-      _ <- fs2.Stream
-            .fromIterator(finalizedBlockSet.grouped(chunkSize.toInt))
-            .evalMapAccumulate(0L)(
-              (processed, chunk) => {
-                val processSoFar = processed + chunk.size.toLong
-                processChunk(chunk) >> Log[F]
-                  .info(s"Finalized blocks recorded: ${processSoFar} of ${finalizedBlockSet.size}.")
-                  .as((processSoFar, ()))
-              }
-            )
-            .compile
-            .drain
-
-    } yield ()
-  }
 }
